+++ src/defaultsettings.cpp
@@ -55,7 +55,11 @@ void set_default_settings(Settings *settings)
 	settings->setDefault("keymap_special1", "KEY_KEY_E");
 	settings->setDefault("keymap_chat", "KEY_KEY_T");
 	settings->setDefault("keymap_cmd", "/");
+#if IRRLICHT_VERSION_10000  >= 10703
+	settings->setDefault("keymap_console", "KEY_OEM_3");
+#else
 	settings->setDefault("keymap_console", "KEY_F10");
+#endif
 	settings->setDefault("keymap_rangeselect", "KEY_KEY_R");
 	settings->setDefault("keymap_freemove", "KEY_KEY_K");
 	settings->setDefault("keymap_fastmove", "KEY_KEY_J");